Refactored to allow for purpose-built profile files
The only things now passed in as command line args are the
port to listen on, and a variable number of .srp paths
(Sauna Radar Profile file.

-p <PORT>
-r /path/to/srp/file
-r /path/to/other/srp/file

Filters and settings such as screen centre, height and also default colours
for each symbol type are all stored in this profile file, along with the
path to the .sct file to load.

As multiple radar displays may now be loaded, F6 and F7 cycle through these.
Started implementing sector file text drawing. At the moment, labels are all
hard-coded as visible until I implement the filter dialog for these.